Rename `uri` module to `url` module
