+++ libs/@blockprotocol/type-system/crate/src/ontology/shared/object/error.rs
@@ -4,7 +4,7 @@ use thiserror::Error;
 use tsify::Tsify;
 
 use crate::{
-    uri::{ParseBaseUrlError, ParseVersionedUrlError},
+    url::{ParseBaseUrlError, ParseVersionedUrlError},
     ParsePropertyTypeReferenceArrayError, ValidationError,
 };
