+++ libs/@blockprotocol/type-system/crate/src/ontology/url/error.rs
