+++ libs/@blockprotocol/type-system/crate/src/ontology/url/wasm.rs
